% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_docx_unreadable}
\alias{err_docx_unreadable}
\title{docx file cannot be read}
\usage{
err_docx_unreadable(path, cause = NULL)
}
\arguments{
\item{path}{File path}

\item{cause}{Underlying condition or message string}
}
\description{
docx file cannot be read
}
