% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_excel_unreadable}
\alias{err_excel_unreadable}
\title{excel file cannot be read}
\usage{
err_excel_unreadable(path, cause = NULL)
}
\arguments{
\item{path}{File path}

\item{cause}{Underlying condition or message string}
}
\description{
excel file cannot be read
}
