% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shiny.R
\name{generate_shiny}
\alias{generate_shiny}
\title{Generate a Shiny application based on user description}
\usage{
generate_shiny(app_description, path)
}
\arguments{
\item{app_description}{A description of the type of Shiny application to generate}

\item{path}{A path to the directory where the application will be saved}
}
\description{
This function sends a user description of a Shiny application to OpenAI and saves the generated code to a specified path.
}
