% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_levels}
\alias{get_levels}
\title{Get all data/indent levels across pages}
\usage{
get_levels(combined)
}
\arguments{
\item{combined}{Output of combine_pages()}
}
\value{
Named list of indent count named by corresponding level
}
\description{
Get all data/indent levels across pages
}
