% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get all unique parameter values across pages}
\usage{
get_parameters(pages)
}
\arguments{
\item{pages}{Output of parse_rtf()}
}
\value{
Character vector of unique parameter values
}
\description{
Get all unique parameter values across pages
}
