% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_timelines}
\alias{get_timelines}
\title{Get all unique timeline values across pages}
\usage{
get_timelines(combined)
}
\arguments{
\item{combined}{Output of combine_pages()}
}
\value{
Character vector of unique timeline values
}
\description{
Get all unique timeline values across pages
}
