% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldapp.R
\name{houdini_oldapp}
\alias{houdini_oldapp}
\title{Return the Houdini Shiny application object}
\usage{
houdini_oldapp()
}
\value{
A Shiny app object.
}
\description{
Called internally by \code{\link{run_app}}. You can also pass the result
directly to \code{shiny::runApp()} if you need more control.
}
