% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{preprocess_csv}
\alias{preprocess_csv}
\title{Preprocess CSV string into a tibble}
\usage{
preprocess_csv(csv_string)
}
\arguments{
\item{csv_string}{A string containing CSV data}
}
\value{
A tibble dataframe with the processed CSV data
}
\description{
Preprocess CSV string into a tibble
}
