% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Houdini shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{arguments passes to \code{\link[shiny]{runApp}} e.g.
\code{port}, \code{launch.browser}.}
}
\value{
Called for its side effects of launching the app
}
\description{
Opens the Houdini app in your default browser
}
