% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{warn_filter_not_found}
\alias{warn_filter_not_found}
\title{Filter not found}
\usage{
warn_filter_not_found(filter_name, filter_type = NULL)
}
\arguments{
\item{filter_name}{The filter that cannot be found}

\item{filter_type}{The filter type (expected to be parameter or timepoint)}
}
\description{
Filter not found
}
