% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_levels}
\alias{filter_levels}
\title{Filter rows by indent level value}
\usage{
filter_levels(combined, levels)
}
\arguments{
\item{combined}{Output of combine_pages()}

\item{levels}{Character vector of level indent titles to include}
}
\description{
Rows with no content are always kept.
}
