% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data function}
\usage{
generate_data(dataset_description)
}
\value{
A tibble dataframe
}
\description{
This function passes a dataset description to an LLM and returns a dataset in tibble dataframe form
}
