% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Houdini-package.R
\docType{package}
\name{Houdini-package}
\alias{Houdini}
\alias{Houdini-package}
\title{Houdini: Automatic Table & Figure Insertion into Data Containing Documents}
\description{
Facilitates the automatic insertion of data tables & figures into documents. Can pull data from .rtf files and .sas7bdat files and will insert them as tables or images into a word doc at a specified bookmark. Bookmarks and filenames must be referenced in a provided excel document as well as any other filtering requirements.
}
\keyword{internal}
