% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Houdini-package.R
\docType{package}
\name{Houdini-package}
\alias{Houdini}
\alias{Houdini-package}
\title{Houdini: Automatic Table & Figure Insertion into Data Containing Documents}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool for the automatic insertion of RTF tables and figures into Word (.docx) documents at specified bookmarks. It parses RTF output (the kind produced by SAS), lets you interactively filter and trim each table with an excel or in app configuration, and writes a fully-formatted Word document with native OOXML tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aarongodwood.github.io/Houdini/}
  \item \url{https://github.com/aarongodwood/Houdini}
}

}
\author{
\strong{Maintainer}: Aaron Godwood \email{godwoodaaron@gmail.com}

Authors:
\itemize{
  \item Valerie Millar \email{v.millar@moonlaketx.com}
}

}
\keyword{internal}
