% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{houdini_pin_folder}
\alias{houdini_pin_folder}
\title{Pins a folder of RTF files to POSIT Connect}
\usage{
houdini_pin_folder(path, name)
}
\arguments{
\item{path}{the file path to the folder of .RTF files to pin}

\item{name}{the name the pin will take on POSIT Connect}
}
\value{
The fully qualified name of the new pin invisibly
}
\description{
Pins a folder of RTF files to POSIT Connect
}
