% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_rtf}
\alias{parse_rtf}
\title{Parse an RTF file into a list of page objects}
\usage{
parse_rtf(path, hide_data = FALSE)
}
\arguments{
\item{path}{Path to the .rtf file}

\item{hide_data}{toggle to replace all data in tables with XX}
}
\value{
List of page objects, each with:
\item{parameter}{character or NA}
\item{header}{block of header rows (see section 4)}
\item{data}{block of data rows, with stable \code{row_id}s}
}
\description{
Parse an RTF file into a list of page objects
}
