% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headless.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read a houdini xlsx workbook}
\usage{
read_xlsx(path)
}
\arguments{
\item{path}{Path to the .xlsx file}
}
\value{
list(config, selections); see [parse_xl()]
}
\description{
Same format as the app's Excel export: see [parse_xl()].
}
